pyproject.toml
[tool.poetry]
name = "wise-alps-dreambooth"
version = "0.1.1"
description = ""
authors = ["wise.water <wise.water@kakaoent.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
kfp = "^2.7.0"
google-cloud-aiplatform = "^1.47.0"
google-cloud-pipeline-components = "^2.13.1"
python-dotenv = "^1.0.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
kfp-kubernetes = "^1.2.0"
diffusers = {git = "https://github.com/huggingface/diffusers"}
accelerate = ">=0.16.0"
torchvision = "^0.17.2"
transformers = ">=4.25.1"
ftfy = "^6.2.0"
tensorboard = "^2.16.2"
jinja2 = "^3.1.3"
peft = "0.7.0"
torch = {version = "^2.2.2+cu121", source = "pytorch"}
torchaudio = {version = "^2.2.2+cu121", source = "pytorch"}
xformers = "^0.0.25.post1"
lightning = "^2.2.2"
torchmetrics = "^1.3.2"
hydra-core = "^1.3.2"
hydra-colorlog = "^1.2.0"
rootutils = "^1.0.7"
pre-commit = "^3.7.0"
rich = "^13.7.1"
wandb = "^0.16.6"
mlflow = "^2.12.1"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
docker = "^6.1.3"
pytest = "^7.4.4"
fire = "^0.5.0"
jinja2 = "^3.1.2"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
